Profile tests: mock out SignedInUserContext
SignedInUserContext needs to be mocked in multiple places. This commit
adds a MockSignedInUserContext that can be used for this purpose.
